Remove pupil class from statistics
The hentInstBruger call needed to get the pupils class name requires
that the school accepts a agreement in STIL. Getting them to do that
seems to be an issue, so it has been decided not to log class names.

As the class wasn't used for much anyway, this isn't that much of a
problem, but we're keeping the existing data around for historical
purposes.